Set duration as a float

+++ bodzify_api/model/track/LibraryTrack.py
@@ -39,7 +39,7 @@ class LibraryTrack(models.Model):
     artist = models.ForeignKey(Artist, on_delete=models.CASCADE, default=None, null=True)
     album = models.ForeignKey(Album, on_delete=models.CASCADE, default=None, null=True)
     genre = models.ForeignKey(Criteria, on_delete=models.DO_NOTHING)
-    duration = models.IntegerField()
+    duration = models.FloatField()
     rating = models.IntegerField(default=0)
     playlists = models.ManyToManyField(Playlist)
     language = models.CharField(max_length=200, default=None, null=True)
